logging: retain annual statistics in memory

+++ iLand/Constant.cs
@@ -4,7 +4,7 @@ namespace iLand
 {
     internal static class Constant
     {
-        public const double AutotrophicRespiration = 0.47;
+        public const float AutotrophicRespiration = 0.47F;
         public const float BiomassCFraction = 0.5F; // fraction of (dry) biomass which is carbon
         public const int DaysInLeapYear = 366;
